Refactor main.c: split into matrix and settings.
